This settings manager branch was a bomb.
We can't make a settings manager run inside the main loop of awesome wm because it makes it not seen as a client by the window manager. The remedy would be to run it as an async process, but that will introduce complexity and bugs to the codebase. Saving branch for reference.